requirements.txt
google-adk>=0.1.0
requests>=2.25.0
